reload when dbfile is updated
postfix will pick up changes automatically when the server is quiescent,
but it never is on the MX-in hosts. this will reload the server config
when spam maps are updated.

this patch also switches the debian restart command to a graceful
reload. there is no need to do a hard restart.